Feat(Checkbox): Allows redering checkboxes

+++ FormService/DTO/DesignDTO.cs
@@ -47,5 +47,5 @@ public class ListItem
 {
     public int list_item_id { get; set; }
     public string name { get; set; }
-    public int numeric_value { get; set; }
+    public double numeric_value { get; set; }
 }
